fix: 챌린지 참여 취소 버그 수정 - 챌린지 참여 상태아닌상태에서 취소 api 호출시 참가자수 1씩 계속 감소하는 문제 해결

+++ backend/src/main/java/com/ssafy/backend/domain/challenge/service/ChallengeServiceImpl.java
@@ -134,12 +134,14 @@ public class ChallengeServiceImpl implements ChallengeService {
     public void leaveChallenge(Long challengeId, Long userId) {
 
         log.info("챌린지 참여 취소 비즈니스 로직 들어왔나?");
+        boolean check = userChallengeRepository.existsByChallenge_ChallengeIdAndUser_UserId(challengeId, userId);
 
-        Challenge challenge = challengeRepository.getReferenceById(challengeId);
-        challenge.subEntry(); // 참가자수 1 감소
+        if (check) {
+            Challenge challenge = challengeRepository.getReferenceById(challengeId);
+            challenge.subEntry(); // 참가자수 1 감소
 
-        userChallengeRepository.deleteByChallenge_ChallengeIdAndUser_UserId(challengeId, userId);
-
+            userChallengeRepository.deleteByChallenge_ChallengeIdAndUser_UserId(challengeId, userId);
+        }
     }
 
 }
